build: Add baseline `pyproject.toml`
This adds the necessary minimum configuration required to build and
import the project. Consequently, this means that we're now able to
install the package in "editable mode" via `pip install -e .`. Any
changes to the source tree will be automatically reflected in the
package; no need to re-install each time.

I've opted for setuptools as the build backend as it's currently what
I'm most familiar with and will be able to do anything we'll need it to.